requirements.txt
Django ~= 1.11
Pillow == 4.2.1
django-multiupload
